perf(state): putting all dashboard modals at the page root
This is to avoid state conflicts and also I had a problem where clicking on a modal nested inside a
component tree would trigger events on the wrapping elements. Which was undesirable. All the
attempts to fix it by preventing the event bubbling up resulted in React Hook Forms not working (I
guess it prevented the onSubmit event). So the best solution is to put the state on the URL, as it's
simple enough for now, and move to better state solutions in the future if they are needed. For now
everything works well again!