feat(documents): fixed document schema
It's more efficient to do it as a BYTEA type, but storing it as string gives the developer much
easier visual feedback on if the document has been uploaded correctly because we can quickly see
what hash the document has in Prisma Studio. And I think the performance impact isn't that high for
the scale of this app.