refactor(prompt): moved default system prompt to a function and out of the database
Getting now the default from the code instead of the database, this way we can update the system
prompt for all conversations with no user defined system prompt, thus, enhancing the app for
everyone with one update. Because prompt engineering can offer benefits to the user. Before, we
would have to update the db for every conversation with some kind of algorithm to detect if the
system prompt was the default one or not. Too complex I think.